Move invalidation tests to stream-invalidate.c
